requirements.txt
streamlit==1.31.0
firebase-admin==6.4.0
python-dotenv==1.0.0
Pillow==10.2.0
plotly==5.18.0
pandas==2.1.4
boto3==1.40.2